+++ Vasilyev/1 lesson (Hello_world!) - 02.09.16/Hello_world!.cpp
