+++ Vasilyev/2 lesson (Calculator) - 09.09.16/Calculator.cpp
