+++ Vasilyev/3 lesson (Max_Int, 3cicl, Check, CheckInt) - 23.09.16/CheckInt.cpp
@@ -1,0 +1,23 @@
+#include "stdafx.h"
+#include <iostream>
+#include <string>
+
+using namespace std;
+
+int main()
+{
+	int n, m, k = 0;
+	string s;
+	cout << "Enter integer : ";
+	cin >> s;
+	cout << endl;
+	try	{
+		n = stoi(s);
+	}
+	catch (...) {
+		cout << "Error";
+		return 0;
+	}
+	cout << "Good!";
+	return 0;
+}
